refactor(local-ci): extract normalize.py for config + priority helpers
R2-1 second cut. Move the 13 config/priority/desktop-adapter
normalization helpers (normalize_priority, priority_value,
normalize_validation_mode, normalize_desktop_*, default_desktop_*,
parse_config_bool, infer_desktop_adapter, normalize_desktop_config,
plus the PRIORITY_VALUES constant) out of local_ci.py into a sibling
normalize.py module.

- local_ci.py: 11,202 → 11,078 lines (-124).
- normalize.py: 164 lines, no behavioral change.
- PRIORITY_VALUES is now owned by normalize.py and re-exported back so
  any consumer reading `local_ci.PRIORITY_VALUES` keeps working.
- The desktop_automation block normalization now lives next to the
  smaller helpers it composes (default_desktop_artifact_root,
  infer_desktop_adapter, default_desktop_bootstrap,
  default_desktop_capability_tier), which were already coupled — this
  just moves the cluster as a unit.

Tests: 213/213 local_ci tests pass.
